apps/api/v1/pyproject.toml
[project]
name = "orchestrator-ai-api-v1"
version = "0.1.0"
description = "Orchestrator AI API - Version 1"
authors = [
    {name = "golfergeek", email = "golfergeet@gmail.com"},
]
dependencies = [
    "fastapi>=0.115.12",
    "PyYAML>=6.0.1",
    "uvicorn[standard]>=0.20.0",
    "python-dotenv>=1.1.0",
    "pydantic[email]>=2.11.4",
    "pydantic-settings>=2.2.1",
    "httpx>=0.27.0",
    "openai>=1.25.2",
    "setuptools>=70.0.0",
    "httpx-sse>=0.4.0",
    "tiktoken>=0.7.0",
    "aiofiles>=23.2.1",
    "python-multipart>=0.0.20",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "supabase>=2.15.1"
]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.26.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.28.1"
]

[tool.pdm]
distribution = false

[tool.setuptools.packages.find]
where = [""]
include = ["main*", "core*", "agents*", "a2a_protocol*", "llm*", "shared*"]
exclude = ["tests*", ".venv*", "__pycache__*"]

[tool.pytest.ini_options]
python_files = "tests.py test_*.py *_tests.py"
addopts = "--asyncio-mode=auto"
asyncio_default_fixture_loop_scope = "function"
